Renaming SubmitButton.jsx to submit_button.jsx for consistency
